# TransportTicketApp/openapi.yaml
openapi: 3.0.3
info:
  version: '1.0'
  title: TransportTicketsMicroservicesApplication
  description: Веб-приложение для предоставления возможностей пользователям покупать
    транспортные билеты
servers:
  - url: http://localhost:8070
    description: Generated server url
paths:
...
  /api/auth/token:
    post:
      tags:
        - API по работе с авторизацией и регистрацией
      operationId: getNewAccessToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshJwtRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /api/auth/register:
    post:
      tags:
        - API по работе с авторизацией и регистрацией
      summary: Регистрация нового пользователя
      description: Нужно заполнить параметры для регистрации
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        "201":
          description: Пользователь зарегистрирован (Created)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterRequest'
        "400":
          description: "Пользователь не зарегистрирован, т.к. не прошел валидацию\
            \ (Bad Request)"
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
        "500":
          description: "Пользователь не зарегистрирован, т.к. такая почта уже суще\
            ствует (Internal Server Error)"
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
  /api/auth/refresh:
    post:
      tags:
        - API по работе с авторизацией и регистрацией
      operationId: getNewRefreshToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshJwtRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /api/auth/login:
    post:
      tags:
        - API по работе с авторизацией и регистрацией
      summary: Авторизация пользователя
      description: Нужно заполнить параметры для авторизации
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        "200":
          description: Пользователь авторизирован (OK)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginRequest'
        "403":
          description: "Пользователь не авторизирован, т.к. ввел неправильные знач\
            ения (Forbidden)"
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoginResponse'
...
  /api/auth/info:
    get:
      tags:
        - API по работе с авторизацией и регистрацией
      summary: Получить информацию о пользователе
      description: Нужно ввести валидный токен
      operationId: getUserInfo
      responses:
        "200":
          description: Данные пользователя получены (OK)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        "500":
          description: "Данные пользователя не получены, т.к. токен не валиден (Internal\
            \ Server Error)"
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
...
components:
  schemas:
...
    RefreshJwtRequest:
      required:
        - refreshToken
      type: object
      properties:
        refreshToken:
          type: string
    LoginResponse:
      type: object
      properties:
        user_id:
          type: integer
          format: int64
        access_token:
          type: string
          description: токен для доступа
        refresh_token:
          type: string
          description: токен для обновления токена для доступа
    RegisterRequest:
      required:
        - email
        - first_name
        - last_name
        - password
      type: object
      properties:
        first_name:
          maxLength: 30
          minLength: 2
          type: string
          description: имя пользователя
        last_name:
          maxLength: 30
          minLength: 2
          type: string
          description: фамилия пользователя
        email:
          type: string
          description: email пользователя
        password:
          maxLength: 30
          minLength: 8
          type: string
          description: пароль пользователя
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
          description: логин пользователя
        password:
          type: string
          description: пароль пользователя
          writeOnly: true
        email:
          type: string
          description: почта пользователя
        role:
          type: string
          description: роль пользователя
          enum:
            - ADMIN
            - USER
        first_name:
          type: string
          description: имя пользователя
        last_name:
          type: string
          description: фамилия пользователя
        enabled:
          type: boolean
          description: доступ пользователя
        created_at:
          type: string
          description: дата регистрации пользователя
          format: date-time
        updated_at:
          type: string
          description: дата изменения информации о пользователе
          format: date-time
    LoginRequest:
      required:
        - password
        - username
      type: object
      properties:
        username:
          maxLength: 30
          minLength: 2
          type: string
          description: логин
        password:
          type: string
          description: пароль
...