# TransportTicketApp/src/main/java/com/github/alekseypetkun/TransportTicketApp/dto/LoginResponse.java
package com.github.alekseypetkun.TransportTicketApp.dto;

import com.fasterxml.jackson.databind.PropertyNamingStrategies;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import io.swagger.v3.oas.annotations.media.Schema;
import lombok.*;

/**
 * DTO получение токенов
 */
@Getter
@Setter
@Builder(toBuilder = true)
@NoArgsConstructor
@AllArgsConstructor
@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
public class LoginResponse {

    private Long userId;

    @Schema(description = "токен для доступа")
    private String accessToken;

    @Schema(description = "токен для обновления токена для доступа")
    private String refreshToken;
}